Revert "Wip/existentialism"
