+++ data_io/__init__.py
@@ -1,8 +1,6 @@
-from . import location
-from . import connection
-from . import data_source
-
-# Export (relevant) builders to expose API
-from .location import build_location
-from .connection import build_connection
-from .data_source import build_data_source
+"""
+For now, just expose the high-level API.
+Bishop wants to reorganize things before other pieces are exposed.
+Exposing those pieces too early will lead to major API issues later.
+"""
+from .data_io import build_connection, connection_builder, location_builder
